Move metrics tracker from src/shared/reporting to plugin
createMetricsTracker was the only export in src/shared/reporting even though it is consumed solely by plugin modules. Move the implementation into src/plugin/src/reporting so plugin metrics stay co-located with their consumers and shared/ no longer exposes plugin-only helpers.
