+++ src/plugin/src/project-index/metrics.js
@@ -1,4 +1,4 @@
-import { createMetricsTracker } from "../../../shared/reporting.js";
+import { createMetricsTracker } from "../reporting/index.js";
 
 const PROJECT_INDEX_METRICS_CATEGORY = "project-index";
 const REQUIRED_METRIC_METHODS = [
