--- src/shared/reporting/index.js
@@ -1,1 +1,0 @@
-export { createMetricsTracker } from "./metrics-tracker.js";
